fix a problem with files > 2GB (thanks to T.J.Adye@rl.ac.uk)

+++ flist.c
@@ -999,12 +999,12 @@ static void clean_flist(struct file_list *flist, int strip_root)
 	if (verbose <= 3) return;
 
 	for (i=0;i<flist->count;i++) {
-		rprintf(FINFO,"[%d] i=%d %s %s mode=0%o len=%d\n",
+		rprintf(FINFO,"[%d] i=%d %s %s mode=0%o len=%.0f\n",
 			getpid(), i, 
 			NS(flist->files[i]->dirname),
 			NS(flist->files[i]->basename),
 			flist->files[i]->mode,
-			(int)flist->files[i]->length);
+			(double)flist->files[i]->length);
 	}
 }
